requirements.txt
base58==2.1.0
importlib-metadata==1.7.0
joblib==0.16.0
lxml==4.6.3
openpyxl==2.5.12
python-docx==0.8.10
pandas
streamlit
xlrd==1.2.0
XlsxWriter==1.1.2
